[refactor] [#46] 게시글 생성 페이지 리팩토링
- 게시글 수정과 생성의 양식이 완전히 동일
- 게시글 생성의 template을 수정에도 동일하게 적용가능함
- 두 개의 페이지를 한 폴더에 넣는 방식으로 구현예정
