+++ src/pages/__test__/CreatePost.test.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import { waitFor } from '@testing-library/react';
 import { render } from '../../test-utils';
-import CreatePost from '../CreatePost';
+import CreatePost from '../ChangePost/CreatePost';
 import { getCategoriesAPI } from '../../api/post';
 
 jest.mock('../../api/post');
